+++ red_utils/utils/__init__.py
@@ -3,6 +3,7 @@ from .stdlib_utils import context_managers
 from .stdlib_utils import dict_utils
 from .stdlib_utils import hash_utils
 from .stdlib_utils import uuid_utils
+from .stdlib_utils import time_utils
 
 # from __future__ import annotations
