+++ red_utils/utils/extended_utils/README.md
@@ -1,0 +1,3 @@
+# extended utils
+
+Utilities in this directory extend 3rd party modules/libraries, such as `SQLAlchemy` or `httpx`.
