+++ red_utils/utils/extended_utils/arrow_utils/constants.py
@@ -1,0 +1,2 @@
+default_format: str = "%Y-%m-%d_%H:%M:%S"
+twelve_hour_format: str = "%Y-%m-%d_%I:%M:%S%p"
