+++ red_utils/utils/extended_utils/arrow_utils/validators.py
@@ -1,0 +1,10 @@
+## List of accepted values for 'period'
+valid_time_periods: list[str] = [
+    "years",
+    "months",
+    "weeks",
+    "days",
+    "hours",
+    "minutes",
+    "seconds",
+]
