+++ red_utils/utils/stdlib_utils/time_utils/__init__.py
@@ -1,4 +1,6 @@
 from __future__ import annotations
 
+import pkgutil
+
 from .constants import default_format, twelve_hour_format
 from .operations import datetime_as_dt, datetime_as_str, get_ts
